upload files from objectstore

+++ bag/objectstore/tests/test_objectstore.py
@@ -1,6 +1,7 @@
 import os
 import mimetypes
 from django.test import TestCase
+
 from ..objectstore  import ObjectStore
 
 
@@ -9,12 +10,12 @@ class TestObjectstore(TestCase):
     def setUp(self):
         self.objectstore = ObjectStore()
 
-    def test_upload(self):
+    def test_objects(self):
 
         res = self.objectstore._get_full_container_list([])
-        self.assertEqual(len(res), 0)
+        self.assertEqual(len(res), 7)
 
-        objects = ['{}'.format(filename) for filename in os.listdir('/bag')]
+        objects = ['diva/bag/{}'.format(filename) for filename in os.listdir('diva/bag')]
         self.assertEqual(len(objects), 29)
 
         for ob in objects:
@@ -23,19 +24,17 @@ class TestObjectstore(TestCase):
             content_type = mimetypes.MimeTypes().guess_type(ob)[0]
             if not content_type:
                 content_type = "application/octet-stream"
-            res = self.objectstore.put_to_objectstore('{}'.format(ob_name), content, content_type)
+            self.objectstore.put_to_objectstore('bagtest/{}'.format(ob_name), content, content_type)
 
         res = self.objectstore._get_full_container_list([])
-        self.assertEqual(len(res), 30)
+        self.assertEqual(len(res), 36)
 
         # clean up
         stored_objects = self.objectstore._get_full_container_list([])
         for ob in stored_objects:
-            self.objectstore.delete_from_objectstore(ob['name'])
+            if ob['name'].startswith('bagtest/'):
+                self.objectstore.delete_from_objectstore(ob['name'])
 
         # check clean up
         res = self.objectstore._get_full_container_list([])
-        self.assertEqual(len(res), 0)
-
-
-
+        self.assertEqual(len(res), 7)
